Python-Test/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
deprecation==2.1.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
postgrest==0.15.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
realtime==1.0.2
six==1.16.0
sniffio==1.3.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.5
supafunc==0.3.3
typing_extensions==4.9.0
websockets==11.0.3
